Conflicts:
	include/clang/Driver/Action.h

+++ tools/clang-offload-bundler/ClangOffloadBundler.cpp
@@ -72,8 +72,21 @@ static cl::opt<bool>
              cl::desc("Unbundle bundled file into several output files.\n"),
              cl::init(false), cl::cat(ClangOffloadBundlerCategory));
 
+static cl::opt<bool> PrintExternalCommands(
+    "###", cl::desc("Print the external commands that are to be executed "
+                    "instead of actually execute them.\n"),
+    cl::init(false), cl::cat(ClangOffloadBundlerCategory));
+
+static cl::opt<bool> SaveTemps("save-temps",
+                               cl::desc("Keep any created temporary files.\n"),
+                               cl::init(false),
+                               cl::cat(ClangOffloadBundlerCategory));
+
 /// Magic string that marks the existence of offloading data.
 #define OFFLOAD_BUNDLER_MAGIC_STR "__CLANG_OFFLOAD_BUNDLE__"
+
+/// The index of the host input in the list of inputs.
+static unsigned HostInputIndex = ~0u;
 
 /// Path to the current binary.
 static std::string BundlerExecutable;
@@ -420,12 +433,14 @@ public:
 
   void WriteHeader(raw_fd_ostream &OS,
                    ArrayRef<std::unique_ptr<MemoryBuffer>> Inputs) {
+    assert(HostInputIndex != ~0u && "Host input index not defined.");
+
     // Record number of inputs.
     NumberOfInputs = Inputs.size();
 
     // Create an LLVM module to have the content we need to bundle.
     auto *M = new Module("clang-offload-bundle", VMContext);
-    M->setTargetTriple(getTriple(TargetNames.front()));
+    M->setTargetTriple(getTriple(TargetNames[HostInputIndex]));
     AuxModule.reset(M);
   }
   void WriteBundleStart(raw_fd_ostream &OS, StringRef TargetTriple) {
@@ -438,15 +453,25 @@ public:
   bool WriteBundleEnd(raw_fd_ostream &OS, StringRef TargetTriple) {
     assert(NumberOfProcessedInputs <= NumberOfInputs &&
            "Processing more inputs that actually exist!");
+    assert(HostInputIndex != ~0u && "Host input index not defined.");
 
     // If this is not the last output, we don't have to do anything.
     if (NumberOfProcessedInputs != NumberOfInputs)
       return false;
 
-    // Create the bitcode file name to write the resulting code to.
+    // Create the bitcode file name to write the resulting code to. Keep it if
+    // save-temps is active.
     SmallString<128> BitcodeFileName;
-    if (sys::fs::createTemporaryFile("clang-offload-bundler", "bc",
-                                     BitcodeFileName)) {
+    if (SaveTemps) {
+      BitcodeFileName = sys::path::filename(OutputFileNames.front());
+      // Save the current file using the host name and the "bc" extension. If
+      // the output already uses that extension, prepend ".tmp".
+      const char *Extension =
+          (sys::path::extension(OutputFileNames.front()) == ".bc") ? "tmp.bc"
+                                                                   : "bc";
+      sys::path::replace_extension(BitcodeFileName, Extension);
+    } else if (sys::fs::createTemporaryFile("clang-offload-bundler", "bc",
+                                            BitcodeFileName)) {
       llvm::errs() << "error: unable to create temporary file.\n";
       return true;
     }
@@ -477,19 +502,26 @@ public:
     // Do the incremental linking. We write to the output file directly. So, we
     // close it and use the name to pass down to clang.
     OS.close();
-    SmallString<128> TargetName = getTriple(TargetNames.front());
+    SmallString<128> TargetName = getTriple(TargetNames[HostInputIndex]);
     const char *ClangArgs[] = {"clang",
                                "-r",
                                "-target",
                                TargetName.c_str(),
                                "-o",
                                OutputFileNames.front().c_str(),
-                               InputFileNames.front().c_str(),
+                               InputFileNames[HostInputIndex].c_str(),
                                BitcodeFileName.c_str(),
                                "-nostdlib",
                                nullptr};
 
-    if (sys::ExecuteAndWait(ClangBinary.get(), ClangArgs)) {
+    // If the user asked for the commands to be printed out, we do that instead
+    // of executing it.
+    if (PrintExternalCommands) {
+      llvm::errs() << "\"" << ClangBinary.get() << "\"";
+      for (unsigned I = 1; ClangArgs[I]; ++I)
+        llvm::errs() << " \"" << ClangArgs[I] << "\"";
+      llvm::errs() << "\n";
+    } else if (sys::ExecuteAndWait(ClangBinary.get(), ClangArgs)) {
       // Remove bitcode file.
       sys::fs::remove(BitcodeFileName);
 
@@ -497,8 +529,9 @@ public:
       return true;
     }
 
-    // Remove bitcode file.
-    sys::fs::remove(BitcodeFileName);
+    // Remove bitcode file if save-temps is not enabled.
+    if (!SaveTemps)
+      sys::fs::remove(BitcodeFileName);
     return false;
   }
   void WriteBundle(raw_fd_ostream &OS, MemoryBuffer &Input) {
@@ -510,10 +543,11 @@ public:
     SectionName += CurrentTriple;
 
     // Create the constant with the content of the section. For the input we are
-    // bundling into (the first input), this is just a place-holder, so a single
+    // bundling into (the host input), this is just a place-holder, so a single
     // byte is sufficient.
+    assert(HostInputIndex != ~0u && "Host input index undefined??");
     Constant *Content;
-    if (NumberOfProcessedInputs == 1) {
+    if (NumberOfProcessedInputs == HostInputIndex + 1) {
       uint8_t Byte[] = {0};
       Content = ConstantDataArray::get(VMContext, Byte);
     } else
@@ -635,8 +669,12 @@ static FileHandler *CreateObjectFileHandler(MemoryBuffer &FirstInput) {
   Expected<std::unique_ptr<Binary>> BinaryOrErr = createBinary(FirstInput);
 
   // Failed to open the input as a known binary. Use the default binary handler.
-  if (!BinaryOrErr)
+  if (!BinaryOrErr) {
+    // We don't really care about the error (we just consume it), if we could
+    // not get a valid device binary object we use the default binary handler.
+    consumeError(BinaryOrErr.takeError());
     return new BinaryFileHandler();
+  }
 
   // We only support regular object files. If this is not an object file,
   // default to the binary handler. The handler will be owned by the client of
@@ -701,9 +739,10 @@ static bool BundleFiles() {
     InputBuffers[Idx++] = std::move(CodeOrErr.get());
   }
 
-  // Get the file handler.
+  // Get the file handler. We use the host buffer as reference.
+  assert(HostInputIndex != ~0u && "Host input index undefined??");
   std::unique_ptr<FileHandler> FH;
-  FH.reset(CreateFileHandler(*InputBuffers.front().get()));
+  FH.reset(CreateFileHandler(*InputBuffers[HostInputIndex].get()));
 
   // Quit if we don't have a handler.
   if (!FH.get())
@@ -877,6 +916,7 @@ int main(int argc, const char **argv) {
 
   // Verify that the offload kinds and triples are known. We also check that we
   // have exactly one host target.
+  unsigned Index = 0u;
   unsigned HostTargetNum = 0u;
   for (StringRef Target : TargetNames) {
     StringRef Kind;
@@ -904,8 +944,13 @@ int main(int argc, const char **argv) {
       llvm::errs() << ".\n";
     }
 
-    if (KindIsValid && Kind == "host")
+    if (KindIsValid && Kind == "host") {
       ++HostTargetNum;
+      // Save the index of the input that refers to the host.
+      HostInputIndex = Index;
+    }
+
+    ++Index;
   }
 
   if (HostTargetNum != 1) {
